Enhancements to Event and Devotional Controllers
Improved the functionality of EventController by sorting events in order of their start time and limiting the output to eight items. Also, made a minor change in ChurchEventResource to include a short_date attribute for better data representation. Additionally, updated the return type in DevotionalController index function for better code clarity and understanding.